// src/modules/DSWExport/types.ts
export type PackageBundle = PackageBundleObject;
/**
 * ID of organization
 */
export type OrganizationID = string;
/**
 * ID of knowledge model
 */
export type KnowledgeModelID = string;
/**
 * Simple semantic version tag, e.g., ``1.0.45``
 */
export type SemVer = string;
/**
 * Unique composed ID for a package
 */
export type PackageID = string;
/**
 * Standard ISO timestamp format, e.g., ``2019-06-09T14:25:58.438Z``
 */
export type Timestamp = string;
/**
 * An event represents a single change in a KM. There are events for each entity type and usually three types of changes: **Add**, **Edit**, and **Delete**. In some cases, also **Move** event is possible.
 */
export type Event =
  | AddKnowledgeModelEvent
  | EditKnowledgeModelEvent
  | AddChapterEvent
  | EditChapterEvent
  | DeleteChapterEvent
  | AddOptionsQuestionEvent
  | AddListQuestionEvent
  | AddItemSelectQuestionEvent
  | AddValueQuestionEvent
  | AddIntegrationQuestionEvent
  | AddMultiChoiceQuestionEvent
  | AddFileQuestionEvent
  | EditOptionsQuestionEvent
  | EditListQuestionEvent
  | EditItemSelectQuestionEvent
  | EditValueQuestionEvent
  | EditIntegrationQuestionEvent
  | EditMultiChoiceQuestionEvent
  | EditFileQuestionEvent
  | DeleteQuestionEvent
  | AddAnswerEvent
  | EditAnswerEvent
  | DeleteAnswerEvent
  | AddExpertEvent
  | EditExpertEvent
  | DeleteExpertEvent
  | AddResourcePageReferenceEvent
  | AddURLReferenceEvent
  | AddCrossReferenceEvent
  | EditResourcePageReferenceEvent
  | EditURLReferenceEvent
  | EditCrossReferenceEvent
  | DeleteReferenceEvent
  | AddTagEvent
  | EditTagEvent
  | DeleteTagEvent
  | AddAPIIntegrationEvent
  | AddIntegrationEvent
  | EditIntegrationEvent
  | EditIntegrationEvent1
  | DeleteIntegrationEvent
  | MoveQuestionEvent
  | MoveAnswerEvent
  | MoveExpertEvent
  | MoveReferenceEvent
  | AddChoiceEvent
  | EditChoiceEvent
  | DeleteChoiceEvent
  | AddMetricEvent
  | EditMetricEvent
  | DeleteMetricEvent
  | AddPhaseEvent
  | EditPhaseEvent
  | DeletePhaseEvent
  | AddResourceCollectionEvent
  | EditResourceCollectionEvent
  | DeleteResourceCollectionEvent
  | AddResourcePageEvent
  | EditResourcePageEvent
  | DeleteResourcePageEvent;
/**
 * UUID version 4 according to `RFC4122 <https://tools.ietf.org/html/rfc4122>`_
 */
export type UUID = string;
export type ListOfKeyValuePairs = KeyValuePair[];
// ...
export interface PackageBundleObject {
  metamodelVersion: number;
  name: string;
  organizationId: OrganizationID;
  kmId: KnowledgeModelID;
  version: SemVer;
  id: PackageID;
  packages: Package[];
  [k: string]: unknown;
}
/**
 * A single KM package that holds changes (or more precisely events) for the specific version.
 */
export interface Package {
  metamodelVersion?: number;
  name?: string;
  organizationId?: OrganizationID;
  kmId?: KnowledgeModelID;
  version?: SemVer;
  id?: PackageID;
  description?: string;
  readme?: string;
  license?: string;
  forkOfPackageId?: PackageID;
  mergeCheckpointPackageId?: PackageID;
  previousPackageId?: PackageID;
  createdAt?: Timestamp;
  events?: Event[];
  [k: string]: unknown;
}
/**
 * Add a new Knowledge Model (typically only once and the very first event) as the top-level entity
 */
export interface AddKnowledgeModelEvent {
  eventType: "AddKnowledgeModelEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  annotations: ListOfKeyValuePairs;
}
export interface KeyValuePair {
  key?: string;
  value?: string;
  [k: string]: unknown;
}
/**
 * Edit properties of the top-level entity
 */
export interface EditKnowledgeModelEvent {
  eventType: "EditKnowledgeModelEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  chapterUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  tagUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  integrationUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  metricUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  phaseUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  resourcePageCollectionUuids?: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  annotations: {
    changed: boolean;
    value?: ListOfKeyValuePairs;
    [k: string]: unknown;
  };
}
/**
 * Add a new chapter to the KM
 */
export interface AddChapterEvent {
  eventType: "AddChapterEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  title: string;
  text: OptionalString;
  annotations: ListOfKeyValuePairs;
}
/**
 * Edit an existing chapter in the KM
 */
export interface EditChapterEvent {
  eventType: "EditChapterEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  title: {
    changed: boolean;
    value?: string;
    [k: string]: unknown;
  };
  text: {
    changed: boolean;
    value?: OptionalString;
    [k: string]: unknown;
  };
  questionUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  annotations: {
    changed: boolean;
    value?: ListOfKeyValuePairs;
    [k: string]: unknown;
  };
}
/**
 * Delete an existing chapter from the KM
 */
export interface DeleteChapterEvent {
  eventType: "DeleteChapterEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
}
/**
 * Add a new options question to the KM
 */
export interface AddOptionsQuestionEvent {
  eventType: "AddQuestionEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  questionType: "OptionsQuestion";
  title: string;
  text: OptionalString;
  requiredPhaseUuid: DesirabilityPhase;
  tagUuids: UUID[];
  annotations: ListOfKeyValuePairs;
}
/**
 * Add a new list question to the KM
 */
export interface AddListQuestionEvent {
  eventType: "AddQuestionEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  questionType: "ListQuestion";
  title: string;
  text: OptionalString;
  requiredPhaseUuid: DesirabilityPhase;
  tagUuids: UUID[];
  annotations: ListOfKeyValuePairs;
}
/**
 * Add a new item select question to the KM
 */
export interface AddItemSelectQuestionEvent {
  eventType: "AddQuestionEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  listQuestionUuid: UUID;
  createdAt: string;
  questionType: "ListQuestion";
  title: string;
  text: OptionalString;
  requiredPhaseUuid: DesirabilityPhase;
  tagUuids: UUID[];
  annotations: ListOfKeyValuePairs;
}
/**
 * Add a new value question to the KM
 */
export interface AddValueQuestionEvent {
  eventType: "AddQuestionEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  questionType: "ValueQuestion";
  title: string;
  text: OptionalString;
  requiredPhaseUuid: DesirabilityPhase;
  tagUuids: UUID[];
  valueType: ValueType;
  validations: QuestionValidation[];
  annotations: ListOfKeyValuePairs;
}
// ...
export interface AddIntegrationQuestionEvent {
  eventType: "AddQuestionEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  questionType: "IntegrationQuestion";
  title: string;
  text: OptionalString;
  requiredPhaseUuid: DesirabilityPhase;
  tagUuids: UUID[];
  integrationUuid: UUID;
  props: StringMap;
  annotations: ListOfKeyValuePairs;
}
/**
 * Representation of simple map where both keys and values are strings
 */
export interface StringMap {
  /**
   * This interface was referenced by `StringMap`'s JSON-Schema definition
   * via the `patternProperty` ".*".
   */
  [k: string]: string;
}
/**
 * Add a new multi-choice question to the KM
 */
export interface AddMultiChoiceQuestionEvent {
  eventType: "AddQuestionEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  questionType: "MultiChoiceQuestion";
  title: string;
  text: OptionalString;
  requiredPhaseUuid: DesirabilityPhase;
  tagUuids: UUID[];
  annotations: ListOfKeyValuePairs;
}
/**
 * Add a new file question to the KM
 */
export interface AddFileQuestionEvent {
  eventType: "AddQuestionEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  questionType: "FileQuestion";
  title: string;
  text: OptionalString;
  fileTypes: OptionalString;
  maxSize: OptionalInteger;
  requiredPhaseUuid: DesirabilityPhase;
  tagUuids: UUID[];
  annotations: ListOfKeyValuePairs;
}
/**
 * Edit an existing options question in the KM
 */
export interface EditOptionsQuestionEvent {
  eventType: "EditQuestionEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  questionType: "OptionsQuestion";
  title: {
    changed: boolean;
    value?: string;
    [k: string]: unknown;
  };
  text: {
    changed: boolean;
    value?: OptionalString;
    [k: string]: unknown;
  };
  requiredPhaseUuid: {
    changed: boolean;
    value?: DesirabilityPhase;
    [k: string]: unknown;
  };
  tagUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  expertUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  referenceUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  answerUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  annotations: {
    changed: boolean;
    value?: ListOfKeyValuePairs;
    [k: string]: unknown;
  };
}
/**
 * Edit an existing list question in the KM
 */
export interface EditListQuestionEvent {
  eventType: "EditQuestionEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  questionType: "ListQuestion";
  title: {
    changed: boolean;
    value?: string;
    [k: string]: unknown;
  };
  text: {
    changed: boolean;
    value?: OptionalString;
    [k: string]: unknown;
  };
  requiredPhaseUuid: {
    changed: boolean;
    value?: DesirabilityPhase;
    [k: string]: unknown;
  };
  tagUuids?: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  expertUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  referenceUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  itemTemplateQuestionUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  annotations: {
    changed: boolean;
    value?: ListOfKeyValuePairs;
    [k: string]: unknown;
  };
}
/**
 * Edit an existing list question in the KM
 */
export interface EditItemSelectQuestionEvent {
  eventType: "EditQuestionEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  questionType: "ItemSelectQuestion";
  title: {
    changed: boolean;
    value?: string;
    [k: string]: unknown;
  };
  text: {
    changed: boolean;
    value?: OptionalString;
    [k: string]: unknown;
  };
  requiredPhaseUuid: {
    changed: boolean;
    value?: DesirabilityPhase;
    [k: string]: unknown;
  };
  tagUuids?: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  expertUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  referenceUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  listQuestionUuid: {
    changed: boolean;
    value?: UUID;
    [k: string]: unknown;
  };
  annotations: {
    changed: boolean;
    value?: ListOfKeyValuePairs;
    [k: string]: unknown;
  };
}
/**
 * Edit an existing value question in the KM
 */
export interface EditValueQuestionEvent {
  eventType: "EditQuestionEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  questionType: "ValueQuestion";
  title: {
    changed: boolean;
    value?: string;
    [k: string]: unknown;
  };
  text: {
    changed: boolean;
    value?: OptionalString;
    [k: string]: unknown;
  };
  requiredPhaseUuid: {
    changed: boolean;
    value?: DesirabilityPhase;
    [k: string]: unknown;
  };
  tagUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  expertUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  referenceUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  validations: {
    changed: boolean;
    value?: QuestionValidation[];
    [k: string]: unknown;
  };
  valueType: {
    changed: boolean;
    value?: ValueType;
    [k: string]: unknown;
  };
  annotations: {
    changed: boolean;
    value?: ListOfKeyValuePairs;
    [k: string]: unknown;
  };
}
/**
 * Edit an existing integration question in the KM
 */
export interface EditIntegrationQuestionEvent {
  eventType: "EditQuestionEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
  questionType: "IntegrationQuestion";
  title: {
    changed: boolean;
    value?: string;
    [k: string]: unknown;
  };
  text: {
    changed: boolean;
    value?: OptionalString;
    [k: string]: unknown;
  };
  requiredPhaseUuid: {
    changed: boolean;
    value?: DesirabilityPhase;
    [k: string]: unknown;
  };
  tagUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  expertUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  referenceUuids: {
    changed: boolean;
    value?: UUID[];
    [k: string]: unknown;
  };
  integrationUuid: {
    changed: boolean;
    value?: UUID;
    [k: string]: unknown;
  };
  props: {
    changed: boolean;
    value?: StringMap;
    [k: string]: unknown;
  };
  annotations: {
    changed: boolean;
    value?: ListOfKeyValuePairs;
    [k: string]: unknown;
  };
}
// ...
export interface DeleteQuestionEvent {
  eventType: "DeleteQuestionEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
}
// ...
export interface DeleteAnswerEvent {
  eventType: "DeleteAnswerEvent";
  uuid: UUID;
  parentUuid: UUID;
  entityUuid: UUID;
  createdAt: string;
}